Updated suggested views
