+++ apps/admin-x-activitypub/src/hooks/use-activity-pub-queries.ts
@@ -532,15 +532,7 @@ export function useSuggestedProfilesForUser(handle: string, limit = 3) {
     const queryClient = useQueryClient();
     const queryKey = QUERY_KEYS.suggestedProfiles(limit);
 
-    const suggestedHandles = [
-        '@index@activitypub.ghost.org',
-        '@index@john.onolan.org',
-        '@index@www.coffeeandcomplexity.com',
-        '@index@ghost.codenamejimmy.com',
-        '@index@www.syphoncontinuity.com',
-        '@index@www.cosmico.org',
-        '@index@www.russbrown.design'
-    ];
+    const suggestedHandles = Object.values(exploreSites).flatMap(category => category.sites);
 
     const suggestedProfilesQuery = useQuery({
         queryKey,
